+++ changelog.md
@@ -1,12 +1,17 @@
 ## Changelog
 **2.4.1** ([Download](http://downloads.wordpress.org/plugin/uploadcare.2.4.1.zip))
 * Minor bug fix for usage jQuery.
-* Test up to Wordpress 4.7.2
+* Test up to Wordpress 4.7.3
 
 **2.4.0** ([Download](http://downloads.wordpress.org/plugin/uploadcare.2.4.0.zip))
 * Uploadcare widget updated to version 2.6.0
 * Test up to Wordpress 4.4.2
 * update uploadcare-php lib
+
+**2.3.3**
+* Add Welcome screen (with settings)
+* Update Uploadcare-PHP library
+* Uploadcare widget updated to version 2.5.9
 
 **2.3.2** ([Download](http://downloads.wordpress.org/plugin/uploadcare.2.3.2.zip))
 * Fix "add via uploadcare" button in media library